requirements.txt
certifi==2018.8.13
chardet==3.0.4
click==6.7
dlib==19.15.0
face-recognition==1.2.2
face-recognition-models==0.3.0
Flask==1.0.2
future==0.16.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.0
opencv-contrib-python==3.4.2.17
opencv-contrib-python-headless==3.4.2.17
opencv-python==3.4.2.17
opencv-python-headless==3.4.2.17
opencv-utils==0.0.2
Pillow==5.2.0
PyYAML==3.13
requests==2.19.1
six==1.11.0
tabulate==0.8.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
